Placed all classes to separate files.
