+++ src/main/java/core/basesyntax/Bulldozer.java
@@ -1,0 +1,15 @@
+package core.basesyntax;
+
+public class Bulldozer extends Machine {
+
+    @Override
+    public void doWork() {
+        System.out.println("The bulldozer started working");
+    }
+
+    @Override
+    public void stopWork() {
+        System.out.println("The bulldozer stopped working");
+    }
+}
+
