+++ src/main/java/core/basesyntax/Excavator.java
@@ -1,0 +1,14 @@
+package core.basesyntax;
+
+public class Excavator extends Machine {
+
+    @Override
+    public void doWork() {
+        System.out.println("The excavator started working");
+    }
+
+    @Override
+    public void stopWork() {
+        System.out.println("The excavator stopped working");
+    }
+}
