+++ src/main/java/core/basesyntax/Machine.java
@@ -1,2 +1,8 @@
 package core.basesyntax;
 
+abstract class Machine {
+
+    public abstract void doWork();
+
+    public abstract void stopWork();
+}
