+++ src/main/java/core/basesyntax/MainApp.java
@@ -11,52 +11,6 @@ package core.basesyntax;
  * - В классе Main создать добавить по одному екземпляру каждого класса.
  * - Отправить все машини на работу и остановить работу всех машин.
  */
-abstract class Machine {
-
-    public abstract void doWork();
-
-    public abstract void stopWork();
-
-}
-
-class Track extends Machine {
-
-    @Override
-    public void doWork() {
-        System.out.println("The track started working");
-    }
-
-    @Override
-    public void stopWork() {
-        System.out.println("The track stopped working");
-    }
-}
-
-class Bulldozer extends Machine {
-
-    @Override
-    public void doWork() {
-        System.out.println("The bulldozer started working");
-    }
-
-    @Override
-    public void stopWork() {
-        System.out.println("The bulldozer stopped working");
-    }
-}
-
-class Excavator extends Machine {
-
-    @Override
-    public void doWork() {
-        System.out.println("The excavator started working");
-    }
-
-    @Override
-    public void stopWork() {
-        System.out.println("The excavator stopped working");
-    }
-}
 
 public class MainApp {
