+++ src/main/java/core/basesyntax/Track.java
@@ -1,0 +1,14 @@
+package core.basesyntax;
+
+public class Track extends Machine {
+
+    @Override
+    public void doWork() {
+        System.out.println("The track started working");
+    }
+
+    @Override
+    public void stopWork() {
+        System.out.println("The track stopped working");
+    }
+}
